jobs/etl_jobs/internal/export_recommendation/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
aiofiles==24.1.0
    # via cloud-sql-python-connector
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.14
    # via cloud-sql-python-connector
aiosignal==1.3.2
    # via aiohttp
asn1crypto==1.5.1
    # via scramp
async-timeout==5.0.1
    # via aiohttp
attrs==25.3.0
    # via aiohttp
cachetools==5.5.2
    # via google-auth
certifi==2025.1.31
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via typer
cloud-sql-python-connector==1.17.0
    # via -r requirements.in
cryptography==44.0.2
    # via cloud-sql-python-connector
dnspython==2.7.0
    # via cloud-sql-python-connector
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
google-api-core==2.24.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-secret-manager
    #   google-cloud-storage
google-auth==2.38.0
    # via
    #   cloud-sql-python-connector
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-secret-manager
    #   google-cloud-storage
google-cloud-bigquery==3.30.0
    # via -r requirements.in
google-cloud-core==2.4.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-secret-manager==2.23.2
    # via -r requirements.in
google-cloud-storage==3.1.0
    # via -r requirements.in
google-crc32c==1.6.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.69.2
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.2
    # via google-cloud-secret-manager
grpcio==1.71.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.71.0
    # via google-api-core
idna==3.10
    # via
    #   requests
    #   yarl
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
packaging==24.2
    # via google-cloud-bigquery
pg8000==1.31.2
    # via
    #   -r requirements.in
    #   cloud-sql-python-connector
propcache==0.3.0
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-secret-manager
protobuf==5.29.3
    # via
    #   google-api-core
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
psycopg2-binary==2.9.10
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pycparser==2.22
    # via cffi
pygments==2.19.1
    # via rich
python-dateutil==2.9.0.post0
    # via
    #   google-cloud-bigquery
    #   pg8000
requests==2.32.3
    # via
    #   cloud-sql-python-connector
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
rich==13.9.4
    # via typer
rsa==4.9
    # via google-auth
scramp==1.4.5
    # via pg8000
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
tenacity==9.0.0
    # via -r requirements.in
typer==0.15.2
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   multidict
    #   rich
    #   typer
urllib3==2.3.0
    # via requests
yarl==1.18.3
    # via aiohttp
